requirements.txt
# Core dependencies
python-dotenv>=1.0.0  # For environment variable management

# Web scraping
requests>=2.31.0  # For HTTP requests
beautifulsoup4>=4.12.2  # For HTML parsing
lxml>=4.9.3  # XML/HTML parser for BeautifulSoup

# Document processing
python-docx>=1.0.1  # For DOCX file handling
PyMuPDF>=1.23.7  # For PDF processing (also known as fitz)

# Vector database
qdrant-client>=1.6.4  # Python client for Qdrant vector database

# LLM and embedding
langchain>=0.0.335  # Framework for LLM applications
langchain-community>=0.0.1  # Community extensions for LangChain
langchain-google-genai>=0.0.5  # Google Gemini integration for LangChain
google-generativeai>=0.3.1  # Google Generative AI API
sentence-transformers>=2.2.2  # For text embeddings

# Web interface options
flask>=2.3.3  # Web framework option 1
streamlit>=1.28.2  # Web framework option 2 (interactive dashboards)

# Utilities
tqdm>=4.66.1  # Progress bars
pandas>=2.1.3  # Data manipulation
numpy>=1.26.2  # Numerical operations

# Testing
pytest>=7.4.3  # Testing framework

# Ensure specific versions if needed, e.g.:
# langchain==0.1.0
# qdrant-client>=1.7.0